remove import of Base class for test_datetime_index.py
